+++ src/components/store.ts
@@ -1,4 +1,4 @@
 import { writable } from 'svelte/store';
 
-export let team = writable(Array(6).fill(null))
-export let typedex = writable([])
+export const team = writable(Array(6).fill(null))
+export const typedex = writable([])
