+++ src/model/effectiveness.ts
@@ -1,0 +1,8 @@
+export enum Effectiveness {
+    Immune,
+    Quarter,
+    Half,
+    Neutral,
+    Double,
+    Quardrouple
+}
