+++ src/model/pokemon.ts
@@ -1,0 +1,4 @@
+class Pokemon {
+    name: string;
+    types: string[];
+}
