+++ src/model/type.ts
@@ -1,0 +1,6 @@
+export class Type {
+    name: string;
+    noDamageFrom: string[]
+    halfDamageFrom: string[]
+    doubleDamageFrom: string[]
+}
